Move test cases from NodeSettingsTest to IPExtensionsTest
